Add config eample. Cosmetic type fix: unit32 -> uint16 for ports I don't like transport got addresses as strings even we checked it already. Change them onto netip.AddrPort. Fix secret passing.

+++ cmd/main.go
@@ -230,7 +230,7 @@ func main() {
 	// Start processing in a goroutine
 	done := make(chan struct{})
 	go func() {
-		protocol.ResolveAndProcess(ctx, t, host, uint32(port))
+		protocol.ResolveAndProcess(ctx, t, host, uint16(port))
 		close(done)
 	}()
